+++ task-utils.js
@@ -1,3 +1,6 @@
+// helper functions to use from "task.js" plugins code
+// that need access to the file system
+
 // @ts-check
 /// <reference types="cypress" />
 const { readFileSync, writeFileSync, existsSync } = require('fs')
@@ -24,65 +27,6 @@ function checkAllPathsNotFound(nycFilename) {
     allFilesAreMissing
   )
   return allFilesAreMissing
-}
-
-/**
- * remove coverage for the spec files themselves,
- * only keep "external" application source file coverage
- */
-const filterSpecsFromCoverage = (totalCoverage, config = Cypress.config) => {
-  const integrationFolder = config('integrationFolder')
-  // @ts-ignore
-  const testFilePattern = config('testFiles')
-
-  // test files chould be:
-  //  wild card string "**/*.*" (default)
-  //  wild card string "**/*spec.js"
-  //  list of wild card strings or names ["**/*spec.js", "spec-one.js"]
-  const testFilePatterns = Array.isArray(testFilePattern)
-    ? testFilePattern
-    : [testFilePattern]
-
-  const isUsingDefaultTestPattern = testFilePattern === '**/*.*'
-
-  const isTestFile = filename => {
-    const matchedPattern = testFilePatterns.some(specPattern =>
-      Cypress.minimatch(filename, specPattern)
-    )
-    const matchedEndOfPath = testFilePatterns.some(specPattern =>
-      filename.endsWith(specPattern)
-    )
-    return matchedPattern || matchedEndOfPath
-  }
-
-  const isInIntegrationFolder = filename =>
-    filename.startsWith(integrationFolder)
-
-  const isA = (fileCoverge, filename) => isInIntegrationFolder(filename)
-  const isB = (fileCoverge, filename) => isTestFile(filename)
-
-  const isTestFileFilter = isUsingDefaultTestPattern ? isA : isB
-
-  const coverage = Cypress._.omitBy(totalCoverage, isTestFileFilter)
-  return coverage
-}
-
-/**
- * Replace source-map's path by the corresponding absolute file path
- * (coverage report wouldn't work with source-map path being relative
- * or containing Webpack loaders and query parameters)
- */
-function fixSourcePathes(coverage) {
-  Object.values(coverage).forEach(file => {
-    const { path: absolutePath, inputSourceMap } = file
-    const fileName = /([^\/\\]+)$/.exec(absolutePath)[1]
-    if (!inputSourceMap || !fileName) return
-
-    if (inputSourceMap.sourceRoot) inputSourceMap.sourceRoot = ''
-    inputSourceMap.sources = inputSourceMap.sources.map(source =>
-      source.includes(fileName) ? absolutePath : source
-    )
-  })
 }
 
 /**
@@ -250,11 +194,28 @@ function tryFindingLocalFiles(nycFilename) {
   }
 }
 
+/**
+ * Replace source-map's path by the corresponding absolute file path
+ * (coverage report wouldn't work with source-map path being relative
+ * or containing Webpack loaders and query parameters)
+ */
+function fixSourcePaths(coverage) {
+  Object.values(coverage).forEach(file => {
+    const { path: absolutePath, inputSourceMap } = file
+    const fileName = /([^\/\\]+)$/.exec(absolutePath)[1]
+    if (!inputSourceMap || !fileName) return
+
+    if (inputSourceMap.sourceRoot) inputSourceMap.sourceRoot = ''
+    inputSourceMap.sources = inputSourceMap.sources.map(source =>
+      source.includes(fileName) ? absolutePath : source
+    )
+  })
+}
+
 module.exports = {
-  fixSourcePathes,
-  filterSpecsFromCoverage,
   showNycInfo,
   resolveRelativePaths,
   checkAllPathsNotFound,
-  tryFindingLocalFiles
+  tryFindingLocalFiles,
+  fixSourcePaths
 }
